divide stratum view index functions to files
